+++ NEWS.md
@@ -1,5 +1,6 @@
 # PLNmodels 0.10.2 - medium release
 
+* Fix bug in predict() methods when factor levels differ between train and test datasets. 
 * Fix bug in PLNPCAfit S3 plot() method
 * Some simplification in C++ code
 * correction/changes in PLN likelihoods? + added constant terms in all likelihoods of all PLN models